# specs/001-sitemap-monitor/contracts/api.yaml
openapi: 3.1.0
info:
  title: Sitemap Monitor API
  description: API for monitoring sitemap changes
  version: 1.0.0
  contact:
    name: Sitemap Monitor Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: 用户注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 邮箱已存在或参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          headers:
            Set-Cookie:
              description: Refresh token in HTTPOnly cookie
              schema:
                type: string
        '401':
          description: 邮箱或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [auth]
      summary: 用户登出
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 登出成功
  /auth/refresh:
    post:
      tags: [auth]
      summary: 刷新访问令牌
      operationId: refreshToken
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh token 无效或过期
  /auth/password/reset-request:
    post:
      tags: [auth]
      summary: 请求重置密码
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 如果邮箱存在，将发送重置链接
  /auth/password/reset:
    post:
      tags: [auth]
      summary: 重置密码
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 密码重置成功
        '400':
          description: Token 无效或过期

  # ============ 用户 ============
...
  /monitors/{monitor_id}/changes:
    get:
      tags: [changes]
      summary: 获取变更历史
      operationId: listChanges
      security:
        - bearerAuth: []
      parameters:
        - name: monitor_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: change_type
          in: query
          schema:
            $ref: '#/components/schemas/ChangeType'
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeListResponse'
  /monitors/{monitor_id}/changes/{change_id}:
    get:
      tags: [changes]
      summary: 获取变更详情
      operationId: getChange
      security:
        - bearerAuth: []
      parameters:
        - name: monitor_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: change_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeDetailResponse'

  # ============ 通知配置 ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: 过期时间（秒）
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_verified:
          type: boolean
        has_completed_onboarding:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ChangeListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChangeSummaryResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    ChangeSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        change_type:
          $ref: '#/components/schemas/ChangeType'
        added_count:
          type: integer
        removed_count:
          type: integer
        modified_count:
          type: integer
        created_at:
          type: string
          format: date-time
    ChangeDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        change_type:
          $ref: '#/components/schemas/ChangeType'
        added_count:
          type: integer
        removed_count:
          type: integer
        modified_count:
          type: integer
        changes:
          type: object
          properties:
            added:
              type: array
              items:
                $ref: '#/components/schemas/UrlChange'
            removed:
              type: array
              items:
                $ref: '#/components/schemas/UrlChange'
            modified:
              type: array
              items:
                $ref: '#/components/schemas/UrlModification'
        created_at:
          type: string
          format: date-time
    UrlChange:
      type: object
      properties:
        url:
          type: string
          format: uri
        lastmod:
          type: string
          nullable: true
    UrlModification:
      type: object
      properties:
        url:
          type: string
          format: uri
        old_lastmod:
          type: string
          nullable: true
        new_lastmod:
          type: string
          nullable: true
...
    ChangeType:
      type: string
      enum: [initial, no_change, changed]
...